Updated eyed3-display-plugin.py to be compatible with pipx and Python 3.12+
With a new policy Python scripts are now installed using pipx in separated virtual enviroments. This leads to the change in the shebang of the eyeD3 display plugin emulatotor script. The global installation of eyeD3 is assumed here i.e. using the --global flag in pipx install, with the default paths used by pipx.

Moreover in Python 3.12 the error message of the "invalid escape sequence" error is changed from a DeprecationWarning to a SyntaxWarning. That is why the backslashes in the script have to be replaced by the double ones so they properly represent the atcual backslashes in a python interpreter and not raising a syntax error.